requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
pydantic==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
huggingface-hub==0.23.0
requests==2.31.0
pillow==10.3.0
pandas==2.2.2
matplotlib==3.8.4